backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
google-generativeai
websockets
python-dotenv
pydantic
langchain-groq
